requirements.txt
attrs==20.3.0
cachelib==0.1.1
certifi==2020.12.5
click==7.1.2
coverage==5.3
cycler==0.10.0
dataclasses==0.8
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Session==0.3.2
future==0.18.2
importlib-metadata==3.1.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.2
natsort==7.1.0
numpy==1.19.4
packaging==20.7
pandas==1.1.3
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tqdm==4.50.2
typing-extensions==3.7.4.3
visitor==0.1.3
Werkzeug==1.0.1
zipp==3.4.0
